requirements.txt
slack-bolt>=1.21.0
google-generativeai>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
